Basic collision detection, removed GameController
Not added actual movement, but the player position does check to see if
it can move then update it's direction (not fully fixed yet, moves every
frame)

+++ Lab1/MyGame.h
@@ -13,5 +13,5 @@ private:
 
         static const int width = 17;
 		static const int height = 21;
-		byte gridSpots[];
+		int gridSpots[];
 };
